requirements.txt
flask
flask-socketio
flask-login
flask-session
werkzeug
ultralytics
torch
torchvision
opencv-python
numpy
python-dotenv
tqdm
pyyaml
Pillow
# Валидация
pydantic>=2.0.0
pydantic[email]
# Дополнительные утилиты
python-multipart
email-validator
# Аугментация данных
albumentations